allow case insensitive search
I'm using the `FindAllIndex` method for the `SuffixArray` of the complete works to allow the use of a regular expression. Since it returns start and end indixes, I want to ensure the results are still the same, so I'm only using the start index.
I create the regex using MustCompile for brevity, I need to investigate if this could cause issues of not-nicely formatted strings.